+++ utility/module.go
@@ -1,6 +1,8 @@
 package utility
 
 import (
+	"encoding/json"
+	"io/ioutil"
 	"log"
 
 	"github.com/pokt-network/pocket/utility/types"
@@ -21,14 +23,31 @@ const (
 )
 
 func Create(configPath, genesisPath string) (modules.UtilityModule, error) {
+	u := new(UtilityModule)
+	c, err := u.InitConfig(configPath)
+	if err != nil {
+		return nil, err
+	}
+	config := (c).(*types.UtilityConfig)
 	return &UtilityModule{
-		// TODO: Add `maxTransactionBytes` and `maxTransactions` to cfg.Utility
-		Mempool: types.NewMempool(1000, 1000),
+		Mempool: types.NewMempool(config.Max_Mempool_Transaction_Bytes, config.Max_Mempool_Transactions),
 	}, nil
 }
 
 func (u *UtilityModule) InitConfig(pathToConfigJSON string) (config modules.IConfig, err error) {
-	return // No-op
+	data, err := ioutil.ReadFile(pathToConfigJSON)
+	if err != nil {
+		return
+	}
+	// over arching configuration file
+	rawJSON := make(map[string]json.RawMessage)
+	if err = json.Unmarshal(data, &rawJSON); err != nil {
+		log.Fatalf("[ERROR] an error occurred unmarshalling the %s file: %v", pathToConfigJSON, err.Error())
+	}
+	// persistence specific configuration file
+	config = new(types.UtilityConfig)
+	err = json.Unmarshal(rawJSON[u.GetModuleName()], config)
+	return
 }
 
 func (u *UtilityModule) InitGenesis(pathToGenesisJSON string) (genesis modules.IGenesis, err error) {