+++ utility/doc/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.4] - 2022-09-23
+- Created UtilityConfig
+- Added `Max_Mempool_Transaction_Bytes` and `Max_Mempool_Transactions` to the utility 
+  config to allow dynamic configuration of the mempool
+- Matched configuration unmarshalling pattern of other modules
+- Added V0 mempool default configurations
+- Regenerated build files with new mempool config
+
 ## [0.0.0.3] - 2022-09-15
 
 ### Code cleanup